+++ packages/typespec-autorest/test/arm/arm.test.ts
@@ -163,6 +163,58 @@ it("can use private endpoints with common-types references", async () => {
   ok(openapi.paths?.[privateEndpointGet]?.get?.parameters?.[1]);
 });
 
+it("verify resolution of private endpoints and private links with v5 version", async () => {
+  const openapi = await openApiFor(
+    `@useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
+      @armProviderNamespace
+      @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v5)
+      namespace Microsoft.PrivateLinkTest;
+      
+      interface Operations extends Azure.ResourceManager.Operations {}
+      
+      @tenantResource
+      model PrivateEndpointConnectionResource is ProxyResource<PrivateEndpointConnectionProperties> {
+        @path
+        @segment("privateEndpointConnections")
+        @key("privateEndpointConnectionName")
+        name: string;
+      }
+      
+      @armResourceOperations(PrivateEndpointConnectionResource)
+      interface PrivateEndpointConnections {
+        #suppress "deprecated" "PrivateLinkResourceListResultV5 validation"
+        listConnections is ArmResourceListByParent<PrivateEndpointConnectionResource,
+         Response = ArmResponse<Azure.ResourceManager.CommonTypes.PrivateEndpointConnectionListResultV5>>;
+      }
+
+      model PrivateLinkResource is ProxyResource<PrivateLinkResourceProperties> {
+        ...PrivateLinkResourceParameter;
+      }
+
+      @armResourceOperations(PrivateLinkResource)
+      interface PrivateLinkResources {
+        #suppress "deprecated" "PrivateLinkResourceListResultV5 validation"
+        listByLinkResult is ArmResourceListByParent< PrivateLinkResource,
+          Response = ArmResponse<Azure.ResourceManager.CommonTypes.PrivateLinkResourceListResultV5>
+        >;
+      }
+      `,
+  );
+
+  const privateEndpointList = "/providers/Microsoft.PrivateLinkTest/privateEndpointConnections";
+  const privateLinkList =
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.PrivateLinkTest/privateLinkResources";
+
+  deepStrictEqual(
+    openapi.paths[privateEndpointList].get.responses["200"].schema["$ref"],
+    "../../common-types/resource-management/v5/privatelinks.json#/definitions/PrivateEndpointConnectionListResult",
+  );
+  deepStrictEqual(
+    openapi.paths[privateLinkList].get.responses["200"].schema["$ref"],
+    "../../common-types/resource-management/v5/privatelinks.json#/definitions/PrivateLinkResourceListResult",
+  );
+});
+
 it("can use ResourceNameParameter for custom name parameter definition", async () => {
   const openapi: any = await compileOpenAPI(
     `@useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)