+++ packages/typespec-autorest/test/azure-core-types.test.ts
@@ -51,7 +51,7 @@ describe("armResourceIdentifier", () => {
     const result = await compileOpenAPI(
       `
       ${base}
-      scalar Foo extends Azure.Core.armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["tenant", "resourceGroup"]}]>;
+      scalar Foo extends Azure.Core.armResourceIdentifier<[{type:"Microsoft.RP/type", scopes:["Tenant", "ResourceGroup"]}]>;
     `,
       { preset: "azure" },
     );
@@ -63,7 +63,7 @@ describe("armResourceIdentifier", () => {
       "x-ms-arm-id-details": {
         allowedResources: [
           {
-            scopes: ["tenant", "resourceGroup"],
+            scopes: ["Tenant", "ResourceGroup"],
             type: "Microsoft.RP/type",
           },
         ],