+++ packages/typespec-client-generator-core/src/internal-utils.ts
@@ -8,7 +8,7 @@ import {
   getLifecycleVisibilityEnum,
   getNamespaceFullName,
   getVisibilityForClass,
-  Interface,
+  ignoreDiagnostics,
   isNeverType,
   isNullType,
   isVoidType,
@@ -45,6 +45,7 @@ import {
   SdkEnumType,
   SdkHttpResponse,
   SdkModelPropertyType,
+  SdkOperationGroup,
   SdkType,
   TCGCContext,
 } from "./interfaces.js";
@@ -55,7 +56,7 @@ import {
   getHttpOperationWithCache,
   isApiVersion,
 } from "./public-utils.js";
-import { getClientTypeWithDiagnostics } from "./types.js";
+import { getClientTypeWithDiagnostics, getSdkModelPropertyType } from "./types.js";
 
 export interface TCGCEmitterOptions extends BrandedSdkEmitterOptionsInterface {
   "emitter-name"?: string;
@@ -87,21 +88,67 @@ export const negationScopesKey = createStateSymbol("negationScopes");
 export const scopeKey = createStateSymbol("scope");
 export const clientKey = createStateSymbol("client");
 export const operationGroupKey = createStateSymbol("operationGroup");
+export const clientLocationKey = createStateSymbol("clientLocation");
+export const omitOperation = createStateSymbol("omitOperation");
 
 export function hasExplicitClientOrOperationGroup(context: TCGCContext): boolean {
   return (
-    listScopedDecoratorData(context, clientKey).length > 0 ||
-    listScopedDecoratorData(context, operationGroupKey).length > 0
+    listScopedDecoratorData(context, clientKey).size > 0 ||
+    listScopedDecoratorData(context, operationGroupKey).size > 0
   );
 }
 
-function listScopedDecoratorData(context: TCGCContext, key: symbol): any[] {
-  const retval = [...context.program.stateMap(key).values()];
-  return retval
-    .filter((targetEntry) => {
-      return targetEntry[context.emitterName] || targetEntry[AllScopes];
-    })
-    .flatMap((targetEntry) => targetEntry[context.emitterName] ?? targetEntry[AllScopes]);
+export function listScopedDecoratorData(
+  context: TCGCContext,
+  key: symbol,
+  languageScope?: string | typeof AllScopes,
+): Map<Type, any> {
+  const scope = languageScope ?? context.emitterName;
+  const retval: Map<Type, any> = new Map();
+  for (const [type, data] of context.program.stateMap(key).entries()) {
+    if (data[scope]) {
+      // positive scope case
+      retval.set(type, data[scope]);
+    } else if (data[negationScopesKey]) {
+      // negative scope case
+      if (data[negationScopesKey].includes(scope)) {
+        // if the scope is negated, we should not include it
+        continue;
+      } else {
+        // if the scope is not negated, we should include it
+        retval.set(type, data[AllScopes]);
+      }
+    } else if (data[AllScopes]) {
+      // all scopes case
+      retval.set(type, data[AllScopes]);
+    }
+  }
+  return retval;
+}
+
+export function getScopedDecoratorData(
+  context: TCGCContext,
+  key: symbol,
+  target: Type,
+  languageScope?: string | typeof AllScopes,
+): any {
+  const retval: Record<string | symbol, any> = context.program.stateMap(key).get(target);
+  if (retval === undefined) return retval;
+  if (languageScope === AllScopes) {
+    return retval[languageScope];
+  }
+  if (languageScope === undefined || typeof languageScope === "string") {
+    const scope = languageScope ?? context.emitterName;
+    if (Object.keys(retval).includes(scope)) return retval[scope];
+
+    // if the scope is negated, we should return undefined
+    // if the scope is not negated, we should return the value for AllScopes
+    const negationScopes = retval[negationScopesKey];
+    if (negationScopes !== undefined && negationScopes.includes(scope)) {
+      return undefined;
+    }
+  }
+  return retval[AllScopes]; // in this case it applies to all languages
 }
 
 /**
@@ -140,8 +187,8 @@ export function parseEmitterName(
  */
 export function updateWithApiVersionInformation(
   context: TCGCContext,
-  type: { name: string },
-  namespace?: Namespace | Interface,
+  type: ModelProperty,
+  client?: SdkClient | SdkOperationGroup,
 ): {
   isApiVersionParam: boolean;
   clientDefaultValue?: string;
@@ -150,8 +197,8 @@ export function updateWithApiVersionInformation(
   return {
     isApiVersionParam,
     clientDefaultValue:
-      isApiVersionParam && namespace
-        ? context.__clientToApiVersionClientDefaultValue.get(namespace)
+      isApiVersionParam && client
+        ? context.__clientApiVersionDefaultValueCache.get(client)
         : undefined,
   };
 }
@@ -427,11 +474,6 @@ export function isSubscriptionId(context: TCGCContext, parameter: { name: string
   return Boolean(context.arm) && parameter.name === "subscriptionId";
 }
 
-export function getLocationOfOperation(operation: Operation): Namespace | Interface {
-  // have to check interface first, because interfaces are more granular than namespaces
-  return (operation.interface || operation.namespace)!;
-}
-
 export function isNeverOrVoidType(type: Type): boolean {
   return isNeverType(type) || isVoidType(type);
 }
@@ -547,14 +589,14 @@ export function isOnClient(
   operation?: Operation,
   versioning?: boolean,
 ): boolean {
-  const namespace = operation ? getLocationOfOperation(operation) : type.model?.namespace;
+  const client = operation ? context.getClientForOperation(operation) : undefined;
   return (
     isSubscriptionId(context, type) ||
     (isApiVersion(context, type) && versioning) ||
     Boolean(
-      namespace &&
-        context.__clientToParameters
-          .get(namespace)
+      client &&
+        context.__clientParametersCache
+          .get(client)
           ?.find((x) => twoParamsEquivalent(context, x.__raw, type)),
     )
   );
@@ -593,7 +635,7 @@ export function hasNoneVisibility(context: TCGCContext, type: ModelProperty): bo
   return visibility.size === 0;
 }
 
-export function listAllNamespaces(
+function listAllNamespaces(
   context: TCGCContext,
   namespace: Namespace,
   retval?: Namespace[],
@@ -719,4 +761,42 @@ export function getClientDoc(context: TCGCContext, target: Type): string | undef
     }
   }
   return baseDoc;
+}
+
+export function compareModelProperties(
+  context: TCGCContext | undefined,
+  modelPropA: ModelProperty | undefined,
+  modelPropB: ModelProperty | undefined,
+): boolean {
+  if (!modelPropA || !modelPropB) return false;
+  if (modelPropA.name !== modelPropB.name || modelPropA.type !== modelPropB.type) return false;
+  if (!context) return true; // if we don't have a context, we can't further compare the types. Assume true.
+  const sdkA = ignoreDiagnostics(getSdkModelPropertyType(context, modelPropA));
+  const sdkB = ignoreDiagnostics(getSdkModelPropertyType(context, modelPropB));
+  if (sdkA.kind === "method" || sdkB.kind === "method") {
+    // if we're comparing method vs service param, we just need to check the name and type
+    return true;
+  }
+  switch (sdkA.kind) {
+    case "cookie":
+    case "header":
+    case "query":
+    case "path":
+    case "responseheader":
+    case "body":
+      return sdkA.kind === sdkB.kind && sdkA.serializedName === sdkB.serializedName;
+    default:
+      return sdkA.kind === sdkB.kind;
+  }
+}
+
+export function* filterMapValuesIterator<V>(
+  iterator: MapIterator<V>,
+  predicate: (value: V) => boolean,
+): MapIterator<V> {
+  for (const value of iterator) {
+    if (predicate(value)) {
+      yield value;
+    }
+  }
 }