+++ packages/typespec-client-generator-core/src/types.ts
@@ -94,7 +94,6 @@ import {
   getClientDoc,
   getHttpBodySpreadModel,
   getHttpOperationResponseHeaders,
-  getLocationOfOperation,
   getNonNullOptions,
   getNullOption,
   getSdkTypeBaseHelper,
@@ -129,7 +128,7 @@ export function getTypeSpecBuiltInType(
   context: TCGCContext,
   kind: IntrinsicScalarName,
 ): SdkBuiltInType {
-  const global = context.getMutatedGlobalNamespace();
+  const global = context.getMutatedGlobalNamespace(); // since other build in types have been mutated, we need to use the mutated global namespace
   const typeSpecNamespace = global.namespaces!.get("TypeSpec");
   const type = typeSpecNamespace!.scalars.get(kind)!;
 
@@ -1244,7 +1243,7 @@ export function getSdkModelPropertyTypeBase(
     ...updateWithApiVersionInformation(
       context,
       type,
-      operation ? getLocationOfOperation(operation) : undefined,
+      operation ? context.getClientForOperation(operation) : undefined,
     ),
     onClient,
     crossLanguageDefinitionId: getCrossLanguageDefinitionId(context, type, operation),
@@ -1369,7 +1368,7 @@ export function getSdkModelPropertyType(
 
   if (!property) {
     const clientParams = operation
-      ? context.__clientToParameters.get(getLocationOfOperation(operation))
+      ? context.__clientParametersCache.get(context.getClientForOperation(operation))
       : undefined;
     const correspondingClientParams = clientParams?.find((x) =>
       twoParamsEquivalent(context, x.__raw, type),
@@ -1508,22 +1507,7 @@ export function updateUsageOrAccess(
 
   if (!options.skipFirst) {
     if (typeof value === "number") {
-      // usage set
-      if (options.isOverride) {
-        // when a type has usage, it could not be override to narrow usage
-        if (
-          ((type.usage & UsageFlags.Input) > 0 && (value & UsageFlags.Input) === 0) ||
-          ((type.usage & UsageFlags.Output) > 0 && (value & UsageFlags.Output) === 0)
-        ) {
-          diagnostics.add(
-            createDiagnostic({
-              code: "conflict-usage-override",
-              target: type.__raw!,
-            }),
-          );
-          return diagnostics.wrap(undefined);
-        }
-      }
+      // usage set is always additive
       type.usage |= value;
     } else {
       // access set
@@ -1568,7 +1552,10 @@ export function updateUsageOrAccess(
   if (!options.propagation) return diagnostics.wrap(undefined);
   if (type.baseModel) {
     options.ignoreSubTypeStack.push(true);
-    if (context.disableUsageAccessPropagationToBase) {
+    if (
+      context.disableUsageAccessPropagationToBase &&
+      type.baseModel.discriminatorProperty === undefined // For models with discriminators, we should not disable propagation
+    ) {
       options.skipFirst = true;
     }
     diagnostics.pipe(updateUsageOrAccess(context, value, type.baseModel, options));
@@ -1816,6 +1803,14 @@ function updateUsageOverride(context: TCGCContext): [void, readonly Diagnostic[]
     const usageOverride = getUsageOverride(context, sdkType.__raw as any);
     if (usageOverride) {
       diagnostics.pipe(updateUsageOrAccess(context, usageOverride, sdkType, { isOverride: true }));
+      if (usageOverride & UsageFlags.Json) {
+        // if a type has Json usage, then it should have serialization options
+        updateSerializationOptions(context, sdkType, ["application/json"]);
+      }
+      if (usageOverride & UsageFlags.Xml) {
+        // if a type has Xml usage, then it should have serialization options
+        updateSerializationOptions(context, sdkType, ["application/xml"]);
+      }
     }
   }
   return diagnostics.wrap(undefined);